index.go: add property for GpuResource
I discoverd the hard way that after transferring an index to the GPU,
the gpu memory isn't freed when the index is deleted. It seems like a
gpuResource is being allocated and holding the memory. Its necessary to
free the gpuResource in order to free the Gpumemory.

In order to free the gpuResource we will need to save it, and call free
on it. In the next commit I will add the Free functionality to
gpu_bindings.go.